Rename monkey_patch_resources to sigic_resources and update app configuration
