+++ src/sigic_geonode/settings.py
@@ -164,7 +164,9 @@ if LDAP_ENABLED and "geonode_ldap" not in INSTALLED_APPS:
 # https://docs.geonode.org/en/master/advanced/contrib/#configuration
 
 INSTALLED_APPS += (
-    # "sigic_geonode.misc",  # esto de los links ya no va a ir dentro de geonode, debe ser un proyecto aparte
+    # "sigic_geonode.misc", #  esto de los links ya no va a ir dentro de geonode, debe ser un proyecto aparte
+    "sigic_geonode.sigic_auth",
+    "sigic_geonode.sigic_resources",
 )
 
 MIDDLEWARE = [
@@ -214,6 +216,3 @@ REST_FRAMEWORK = {
 }
 
 
-INSTALLED_APPS += (
-    "sigic_geonode.monkey_patch_resources",
-)
