+++ src/sigic_geonode/sigic_resources/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = "sigic_geonode.sigic_resources.apps.MonkeyPatchResourcesConfig"
