+++ src/sigic_geonode/sigic_resources/apps.py
@@ -4,13 +4,13 @@ import logging
 logger = logging.getLogger(__name__)
 
 class MonkeyPatchResourcesConfig(AppConfig):
-    name = "sigic_geonode.monkey_patch_resources"
+    name = "sigic_geonode.sigic_resources"
     verbose_name = "Monkey Patch Resources"
 
     def ready(self):
         try:
             # Importa y ejecuta el parche al iniciar Django
-            import sigic_geonode.monkey_patch_resources.patches
+            import sigic_geonode.sigic_resources.patches
             logger.info("Monkey patch importado correctamente")
         except Exception:
             logger.exception("Falló el monkey patch")
